clockManager, executionWindow autoUpdate, settings fix

+++ src/main/java/Settings/Settings.java
@@ -13,11 +13,11 @@ import Clock.ClockManager;
 public class Settings {
 
     private int CPUs;
-    private int instructionDuration;
+    private double instructionDuration;
     private String planningAlgorithm;
     private ClockManager clockManager;
 
-    public Settings (int CPUs, int instructionDuration, String planningAlgorithm,ClockManager clockManager){
+    public Settings (int CPUs, double instructionDuration, String planningAlgorithm,ClockManager clockManager){
         this.CPUs=CPUs;
         this.instructionDuration=instructionDuration;
         this.planningAlgorithm=planningAlgorithm;
@@ -32,11 +32,11 @@ public class Settings {
         this.CPUs = CPUs;
     }
 
-    public int getInstructionDuration() {
+    public double getInstructionDuration() {
         return instructionDuration;
     }
 
-    public void setInstructionDuration(int instructionDuration) {
+    public void setInstructionDuration(double instructionDuration) {
         this.instructionDuration = instructionDuration;
         clockManager.updateInstructionDuration(instructionDuration);
     }
